backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.32.2
python-dotenv==1.0.0
sentence-transformers==2.6.0
faiss-cpu==1.7.4
langchain==0.3.7
langchain-community==0.3.0
langchain-core==0.3.49
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
PyMuPDF==1.23.8
python-docx==0.8.11
aiohttp==3.9.1
eventlet==0.33.3
packaging==24.2.0
tenacity>=8.1.0,<9.0.0 